requirements.txt

hikari-lightbulb==3.0.0a15
hikari==2.1.0
hikari-miru==4.2.0
asyncache==0.3.1
cachetools==5.5.0
asyncpraw==7.8.1
mplcyberpunk==0.7.5
selenium==4.14.0
aiohttp
asyncpg
APScheduler
colorama
python-dotenv
typer
lavalink-rs==0.13.0
pyparsing
mock
jikanpy
fuzzywuzzy
pyyaml
akinator[async]
python-Levenshtein
types-PyYAML
pandas
seaborn
dataenforce
pytimeparse
types-pytz
colormap
easydev
youtube-search-python
humanize
qrcode
tmdb-python
dacite
tabulate
expiring-dict
selenium_async
colorcet
setuptools
emoji==2.12.1